requirements.txt
fastapi[all]==0.115.0
uvicorn[standard]==0.31.0
aiogram==3.17.0
aiosqlite==0.20.0
python-dotenv==1.0.0